Updated interpolation routine used for tier3 regridding and included doi input for functions.
